LayoutTests/fast/regions/element-region-overset-state.html

<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<style>
#article {
    -webkit-flow-into: article;
    border: 2px solid black;
}
#region_1, #region_2, #region_3 {
    -webkit-flow-from: no_article;
    overflow:visible;
    display:inline-block;
    vertical-align:top;
    margin:10px;
}
.block {
   display:block;
}
#region_1 { height:110px; width:200px; }
#region_2 { height:110px; width:200px; }
#region_3 { height:110px; width:200px; }
</style>
</head>
<body>
<div id="description" class="block"></div>
<div id="article">
    <div>
        <div id="content" style="height: 216px; background-color: green;"></div>
    </div>
</div>

<div id="region_1"></div>
<div id="region_2"></div>
<div id="region_3"></div>

<div id="console" class="block"></div>

<script>
    description("Tests Element.webkitRegionOverset attribute")
    if (window.testRunner) {
        testRunner.dumpAsText();
    }
    function regionOverset(id) {
        var e = document.getElementById(id);
        return e.webkitRegionOverset;
    }
    function flowContent(flowName) {
        var r = document.getElementById("region_1");
        r.style.webkitFlowFrom = flowName;
        r = document.getElementById("region_2");
        r.style.webkitFlowFrom = flowName;
        r = document.getElementById("region_3");
        r.style.webkitFlowFrom = flowName;
    }
    function addVisualOverflow() {
        var c = document.getElementById("article");
        c.style.webkitBoxShadow="0px 30px lime";
    }
    function addMoreContent() {
        var c = document.getElementById("content");
        c.style.height = 326;
    }

    function test() {
        shouldBeEqualToString("regionOverset('region_1')", "empty");
        shouldBeEqualToString("regionOverset('region_2')", "empty");
        shouldBeEqualToString("regionOverset('region_3')", "empty");

        debug("Flow content");
        flowContent("article");

        shouldBeEqualToString("regionOverset('region_1')", "fit");
        shouldBeEqualToString("regionOverset('region_2')", "fit");
        shouldBeEqualToString("regionOverset('region_3')", "empty");

        debug("Add more content");
        addMoreContent();

        shouldBeEqualToString("regionOverset('region_1')", "fit");
        shouldBeEqualToString("regionOverset('region_2')", "fit");
        shouldBeEqualToString("regionOverset('region_3')", "fit");

        debug("Add visual overflow");
        addVisualOverflow();

        shouldBeEqualToString("regionOverset('region_1')", "fit");
        shouldBeEqualToString("regionOverset('region_2')", "fit");
        shouldBeEqualToString("regionOverset('region_3')", "overset");

        debug("Empty content");
        flowContent("no_article");

        shouldBeEqualToString("regionOverset('region_1')", "empty");
        shouldBeEqualToString("regionOverset('region_2')", "empty");
        shouldBeEqualToString("regionOverset('region_3')", "empty");
    }
    test();
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
